Extract from HelpDialog the code related to ACRA
